refactor distinct libs filter
use when calculating localProjects from subprojects
